Close client sockets properly
When client sockets encounter errors (as expected when e.g. the client
goes away) the return value from .SOCK_POLL is negative. Need to turn
this to positive when indexing into the list of client sockets.

The net effect of this bug was that once a client socket started
erroring, calls to service() would return immediately regardless of the
timeout, because the closed client socket was never being removed from
the list. This would cause the CPU to get pounded if service() was
being called in a tight loop.